Break dark mode button into its own component
